+++ demo-v5/gatsby-config.js
@@ -1,0 +1,22 @@
+module.exports = {
+  trailingSlash: 'always',
+  siteMetadata: {
+    title: 'Function test',
+  },
+  plugins: [
+    {
+      resolve: `gatsby-source-filesystem`,
+      options: {
+        path: `${__dirname}/content/posts`,
+        name: `blog`,
+      },
+    },
+    {
+      resolve: `gatsby-transformer-remark`,
+      options: {
+        plugins: [],
+      },
+    },
+    'gatsby-plugin-netlify',
+  ],
+}