+++ plugin/test/v4-e2e-test.sh
@@ -1,0 +1,10 @@
+#!/bin/sh
+set -e
+
+for dir in fixtures/v4/*; do
+    echo Running tests in "$dir"
+    cd "$dir"
+    npm i --no-package-lock --legacy-peer-deps
+    npm run test
+    cd -
+done